+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip51Lists/geohashLists/GeohashListDecryptionCache.kt
@@ -20,57 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip51Lists.geohashLists
 
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip51Lists.PrivateTagArrayEventCache
-import com.vitorpamplona.quartz.nip51Lists.geohashList.GeohashListEvent
-import com.vitorpamplona.quartz.nip51Lists.geohashList.geohashSet
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.mapLatest
-import kotlinx.coroutines.flow.onStart
-
-class GeohashListDecryptionCache(
-    val signer: NostrSigner,
-) {
-    val cachedPrivateLists = PrivateTagArrayEventCache<GeohashListEvent>(signer)
-
-    fun cachedGeohashes(event: GeohashListEvent) = cachedPrivateLists.mergeTagListPrecached(event).geohashSet()
-
-    suspend fun geohashes(event: GeohashListEvent) = cachedPrivateLists.mergeTagList(event).geohashSet()
-
-    fun fastStartValueForGeohashList(note: Note): GeohashListCard {
-        val noteEvent = note.event as? GeohashListEvent
-        return if (noteEvent != null) {
-            GeohashListCard(cachedGeohashes(noteEvent).toList())
-        } else {
-            EmptyGeohashListCard
-        }
-    }
-
-    @OptIn(ExperimentalCoroutinesApi::class)
-    fun observeDecryptedGeohashList(note: Note): Flow<GeohashListCard> =
-        note
-            .flow()
-            .metadata.stateFlow
-            .mapLatest { noteState ->
-                val event = noteState.note.event as? GeohashListEvent
-                GeohashListCard(event?.let { geohashes(it).toList() } ?: emptyList())
-            }.onStart {
-                val event = note.event as? GeohashListEvent
-                if (event != null) {
-                    val list = geohashes(event)
-                    if (list.isNotEmpty()) {
-                        emit(GeohashListCard(list.toList()))
-                    } else {
-                        emit(EmptyGeohashListCard)
-                    }
-                } else {
-                    emit(EmptyGeohashListCard)
-                }
-            }.distinctUntilChanged()
-            .flowOn(Dispatchers.IO)
-}
+typealias GeohashListDecryptionCache = com.vitorpamplona.amethyst.commons.model.nip51Lists.geohashLists.GeohashListDecryptionCache
