+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip51Lists/hashtagLists/HashtagListDecryptionCache.kt
@@ -20,17 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip51Lists.hashtagLists
 
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip51Lists.PrivateTagArrayEventCache
-import com.vitorpamplona.quartz.nip51Lists.hashtagList.HashtagListEvent
-import com.vitorpamplona.quartz.nip51Lists.hashtagList.hashtagSet
-
-class HashtagListDecryptionCache(
-    val signer: NostrSigner,
-) {
-    val cachedPrivateLists = PrivateTagArrayEventCache<HashtagListEvent>(signer)
-
-    fun cachedHashtags(event: HashtagListEvent) = cachedPrivateLists.mergeTagListPrecached(event).hashtagSet()
-
-    suspend fun hashtags(event: HashtagListEvent) = cachedPrivateLists.mergeTagList(event).hashtagSet()
-}
+typealias HashtagListDecryptionCache = com.vitorpamplona.amethyst.commons.model.nip51Lists.hashtagLists.HashtagListDecryptionCache
