+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip51Lists/peopleList/PeopleListDecryptionCache.kt
@@ -20,41 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip51Lists.peopleList
 
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip51Lists.PrivateTagArrayEventCache
-import com.vitorpamplona.quartz.nip51Lists.muteList.mutedUserIdSet
-import com.vitorpamplona.quartz.nip51Lists.muteList.mutedUsers
-import com.vitorpamplona.quartz.nip51Lists.muteList.mutedUsersAndWords
-import com.vitorpamplona.quartz.nip51Lists.muteList.mutedWordSet
-import com.vitorpamplona.quartz.nip51Lists.muteList.mutedWords
-import com.vitorpamplona.quartz.nip51Lists.peopleList.PeopleListEvent
-
-class PeopleListDecryptionCache(
-    val signer: NostrSigner,
-) {
-    val cachedPrivateLists = PrivateTagArrayEventCache<PeopleListEvent>(signer)
-
-    fun cachedUsersAndWords(event: PeopleListEvent) = cachedPrivateLists.mergeTagListPrecached(event).mutedUsersAndWords()
-
-    fun cachedUsers(event: PeopleListEvent) = cachedPrivateLists.mergeTagListPrecached(event).mutedUsers()
-
-    fun cachedUserIdSet(event: PeopleListEvent) = cachedPrivateLists.mergeTagListPrecached(event).mutedUserIdSet()
-
-    fun cachedWords(event: PeopleListEvent) = cachedPrivateLists.mergeTagListPrecached(event).mutedWords()
-
-    fun cachedWordSet(event: PeopleListEvent) = cachedPrivateLists.mergeTagListPrecached(event).mutedWordSet()
-
-    suspend fun usersAndWords(event: PeopleListEvent) = cachedPrivateLists.mergeTagList(event).mutedUsersAndWords()
-
-    suspend fun users(event: PeopleListEvent) = cachedPrivateLists.mergeTagList(event).mutedUsers()
-
-    suspend fun userIdSet(event: PeopleListEvent) = cachedPrivateLists.mergeTagList(event).mutedUserIdSet()
-
-    suspend fun words(event: PeopleListEvent) = cachedPrivateLists.mergeTagList(event).mutedWords()
-
-    suspend fun wordSet(event: PeopleListEvent) = cachedPrivateLists.mergeTagList(event).mutedWordSet()
-
-    suspend fun privateUsers(event: PeopleListEvent) = cachedPrivateLists.privateTags(event)?.mutedUsers() ?: emptyList()
-
-    suspend fun privateUserIdSet(event: PeopleListEvent) = cachedPrivateLists.privateTags(event)?.mutedUserIdSet() ?: emptySet()
-}
+typealias PeopleListDecryptionCache = com.vitorpamplona.amethyst.commons.model.nip51Lists.peopleList.PeopleListDecryptionCache
